+++ src/fns.js
@@ -1,6 +1,6 @@
 //add commas to numbers
 const niceNumber = x => {
-  return x.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',')
+  return x.toString().replace(/\B(?=(?:\d{3})+(?!\d))/g, ',')
 }
 
 const pad = function (str, width, char) {
@@ -16,8 +16,4 @@ const duration = start => {
   return ((Date.now() - start) / 1000).toFixed(2)
 }
 
-module.exports = {
-  niceNumber: niceNumber,
-  pad: pad,
-  duration: duration,
-}
+export { niceNumber, pad, duration }